+++ noseOfYeti/tokeniser/tokeniser.py
@@ -2,18 +2,27 @@ from noseOfYeti.tokeniser.tracker import Tracker
 from noseOfYeti.tokeniser.tokens import Tokens
 
 from tokenize import generate_tokens
+import os
+
+WITH_IT_RETURN_TYPE_ENV_NAME = "NOSE_OF_YETI_IT_RETURN_TYPE"
 
 
 class Tokeniser:
     """Endpoint for tokenising a file"""
 
-    def __init__(self, with_describe_attrs=True):
+    def __init__(self, with_describe_attrs=True, with_it_return_type=None):
+        if with_it_return_type is None:
+            self.with_it_return_type = os.environ.get(
+                WITH_IT_RETURN_TYPE_ENV_NAME, ""
+            ).lower() not in ("", "false", "0")
+        else:
+            self.with_it_return_type = with_it_return_type
         self.with_describe_attrs = with_describe_attrs
 
     def translate(self, readline, result=None, no_imports=None):
         # Tracker to keep track of information as the file is processed
         self.tokens = Tokens()
-        self.tracker = Tracker(result, self.tokens)
+        self.tracker = Tracker(result, self.tokens, with_it_return_type=self.with_it_return_type)
 
         try:
             # Looking at all the tokens